+++ src/main/java/org/pokeherb/orderservice/application/service/dto/request/OrderSearchConditionRequestDto.java
@@ -1,0 +1,13 @@
+package org.pokeherb.orderservice.application.service.dto.request;
+
+import org.pokeherb.orderservice.domain.entity.OrderStatus;
+
+import java.util.UUID;
+
+public record OrderSearchConditionRequestDto(
+        UUID orderUserId,
+        UUID requestVendorId,
+        UUID productId,
+        String productName,
+        OrderStatus status
+) { }
