+++ src/main/java/org/pokeherb/orderservice/domain/OrderQueryDao.java
@@ -1,7 +1,7 @@
 package org.pokeherb.orderservice.domain;
 
 import org.pokeherb.orderservice.application.service.dto.request.OrderSearchConditionRequestDto;
-import org.pokeherb.orderservice.presentation.dto.OrderSummaryResponse;
+import org.pokeherb.orderservice.application.service.dto.response.OrderSummaryResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
