+++ src/main/java/org/pokeherb/orderservice/global/infrastructure/config/QuerydslConfig.java
@@ -1,8 +1,7 @@
-package org.pokeherb.orderservice.global.config;
+package org.pokeherb.orderservice.global.infrastructure.config;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import jakarta.persistence.EntityManager;
-import jakarta.persistence.PersistenceContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
