+++ src/main/java/org/pokeherb/orderservice/infrastructure/persistence/OrderQueryDaoImpl.java
@@ -3,14 +3,15 @@ package org.pokeherb.orderservice.infrastructure.persistence;
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
-import org.pokeherb.orderservice.application.service.dto.OrderSearchCondition;
-import org.pokeherb.orderservice.presentation.dto.OrderSummaryResponse;
+import org.pokeherb.orderservice.application.service.dto.request.OrderSearchConditionRequestDto;
+import org.pokeherb.orderservice.domain.OrderQueryDao;
+import org.pokeherb.orderservice.application.service.dto.response.OrderSummaryResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.support.PageableExecutionUtils;
 import org.springframework.stereotype.Repository;
 
-import static org.pokeherb.orderservice.domain.QOrder.order;
+import static org.pokeherb.orderservice.domain.entity.QOrder.order;
 
 @Repository
 @RequiredArgsConstructor
@@ -19,7 +20,7 @@ public class OrderQueryDaoImpl implements OrderQueryDao {
     private final JPAQueryFactory queryFactory;
 
     @Override
-    public Page<OrderSummaryResponse> search(OrderSearchCondition condition, Pageable pageable) {
+    public Page<OrderSummaryResponse> search(OrderSearchConditionRequestDto condition, Pageable pageable) {
 
         var builder = new BooleanBuilder();
 
@@ -31,6 +32,14 @@ public class OrderQueryDaoImpl implements OrderQueryDao {
 
         if (condition.status() != null)
             builder.and(order.orderStatus.eq(condition.status()));
+
+        if(condition.productName() != null && !condition.productName().isBlank())
+            builder.and(order.productName.containsIgnoreCase(condition.productName()));
+        if(condition.requestVendorId() != null)
+            builder.and(order.receiveVendorId.eq(condition.requestVendorId()));
+
+        // 소프트 딜리트 된거 제외
+        builder.and(order.deletedAt.isNull());
 
         var result = queryFactory
                 .selectFrom(order)
